Module aws_sdk_groundstation::model

Modules

antenna_demod_decode_details

See AntennaDemodDecodeDetails

antenna_downlink_config

See AntennaDownlinkConfig

antenna_downlink_demod_decode_config

See AntennaDownlinkDemodDecodeConfig

antenna_uplink_config

See AntennaUplinkConfig

config_list_item

See ConfigListItem

contact_data

See ContactData

dataflow_detail

See DataflowDetail

dataflow_endpoint

See DataflowEndpoint

dataflow_endpoint_config

See DataflowEndpointConfig

dataflow_endpoint_list_item

See DataflowEndpointListItem

decode_config

See DecodeConfig

demodulation_config

See DemodulationConfig

destination

See Destination

eirp

See Eirp

elevation

See Elevation

endpoint_details

See EndpointDetails

frequency

See Frequency

frequency_bandwidth

See FrequencyBandwidth

ground_station_data

See GroundStationData

mission_profile_list_item

See MissionProfileListItem

s3_recording_config

See S3RecordingConfig

s3_recording_details

See S3RecordingDetails

satellite_list_item

See SatelliteListItem

security_details

See SecurityDetails

socket_address

See SocketAddress

source

See Source

spectrum_config

See SpectrumConfig

tracking_config

See TrackingConfig

uplink_echo_config

See UplinkEchoConfig

uplink_spectrum_config

See UplinkSpectrumConfig

Structs

AntennaDemodDecodeDetails

Details about an antenna demod decode Config used in a contact.

AntennaDownlinkConfig

Information about how AWS Ground Station should configure an

AntennaDownlinkDemodDecodeConfig

Information about how AWS Ground Station should conﬁgure an antenna for downlink demod decode during a contact.

AntennaUplinkConfig

Information about the uplink Config of an antenna.

ConfigListItem

An item in a list of Config objects.

ContactData

Data describing a contact.

DataflowDetail

Information about a dataflow edge used in a contact.

DataflowEndpoint

Information about a dataflow endpoint.

DataflowEndpointConfig

Information about the dataflow endpoint Config.

DataflowEndpointListItem

Item in a list of DataflowEndpoint groups.

DecodeConfig

Information about the decode Config.

DemodulationConfig

Information about the demodulation Config.

Destination

Dataflow details for the destination side.

Eirp

Object that represents EIRP.

Elevation

Elevation angle of the satellite in the sky during a contact.

EndpointDetails

Information about the endpoint details.

Frequency

Object that describes the frequency.

FrequencyBandwidth

Object that describes the frequency bandwidth.

GroundStationData

Information about the ground station data.

MissionProfileListItem

Item in a list of mission profiles.

S3RecordingConfig

Information about an S3 recording Config.

S3RecordingDetails

Details about an S3 recording Config used in a contact.

SatelliteListItem

Item in a list of satellites.

SecurityDetails

Information about endpoints.

SocketAddress

Information about the socket address.

Source

Dataflow details for the source side.

SpectrumConfig

Object that describes a spectral Config.

TrackingConfig

Object that determines whether tracking should be used during a contact

UplinkEchoConfig

Information about an uplink echo Config.

UplinkSpectrumConfig

Information about the uplink spectral Config.

Enums

AngleUnits

BandwidthUnits

ConfigCapabilityType

ConfigDetails

ConfigTypeData

ContactStatus

Criticality

EirpUnits

EndpointStatus

FrequencyUnits

Polarization